ConsoleSniffer v3
This is the version 3 cracked by Joshua. It has been around for some time now but it is still working, I can not promise all PAID program cracks are clean your antivirus may or may not pop up, I upload what i use and if it works for me it may just work for you. Also this will need Win10pcap Installed for it to work correctly. This is not my crack but i use it so i do not see why you can't either :) Enjoy as for the console Sniffer v4 its still being worked on and also OctoSniffer